Segfault is fixed; put_img_to_win added; Some mistake is still in the calculation, need to be fixed

+++ srcs/graphic/graphic_utils.c
@@ -6,13 +6,13 @@
 /*   By: khuk <khuk@student.42vienna.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/09 13:11:40 by khuk              #+#    #+#             */
-/*   Updated: 2025/02/14 21:07:04 by khuk             ###   ########.fr       */
+/*   Updated: 2025/02/15 22:32:41 by khuk             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "graphic.h"
 
-int	ft_abs(int i)
+double	ft_abs(double i)
 {
 	int	j;
